Move page object code under test-support
This change prevents the addon code from being included in production
builds.

Nesting the Page Object code under `test-support/` will cause the code
to be merged into the `tests/` directory of the consuming app. When the
consuming app creates a production build, the `tests/` directory is not
included in the build (unless the build was created by running `ember
test`).
